use separate module to package spark-connector
